Drop legacy optimizer
