+++ tools/src/main/scala/scala/scalanative/codegen/VirtualTable.scala
@@ -4,7 +4,6 @@ package codegen
 import scala.collection.mutable
 import scalanative.nir._
 import scalanative.nir.Rt.{Type => _, _}
-import scalanative.sema._
 
 class VirtualTable(meta: Metadata, cls: linker.Class) {
   private val slots: mutable.UnrolledBuffer[Sig] =
