--- tools/src/main/scala/scala/scalanative/optimizer/Pass.scala
@@ -1,7 +1,0 @@
-package scala.scalanative
-package optimizer
-
-trait Pass {
-
-  def onInsts(insts: Seq[nir.Inst]): Seq[nir.Inst]
-}
