--- tools/src/main/scala/scala/scalanative/optimizer/PassCompanion.scala
@@ -1,8 +1,0 @@
-package scala.scalanative
-package optimizer
-
-trait PassCompanion {
-
-  /** Instantiate the given pass. */
-  def apply(config: build.Config, linked: linker.Result): Pass
-}
