+++ tools/src/test/scala/scala/scalanative/codegen/CodeGenSpec.scala
@@ -3,7 +3,6 @@ package codegen
 
 import java.nio.file.{Path, Paths}
 import scalanative.io.VirtualDirectory
-import scalanative.optimizer.Driver
 import scalanative.build.{Config, ScalaNative}
 import scalanative.util.Scope
 
@@ -17,14 +16,11 @@ abstract class CodeGenSpec extends OptimizerSpec {
    * @param sources Map from file name to file content representing all the code
    *                to compile
    * @param fn      A function to apply to the products of the compilation.
-   * @param driver  The driver that defines the pipeline.
    * @return The result of applying `fn` to the resulting file.
    */
-  def codegen[T](
-      entry: String,
-      sources: Map[String, String],
-      driver: Option[Driver] = None)(f: (Config, linker.Result, Path) => T): T =
-    optimize(entry, sources, driver) {
+  def codegen[T](entry: String, sources: Map[String, String])(
+      f: (Config, linker.Result, Path) => T): T =
+    optimize(entry, sources) {
       case (config, optimized) =>
         Scope { implicit in =>
           ScalaNative.codegen(config, optimized)
